Refactored and removed unwanted variables and methods
